Move yelp helpers into separate file
