+++ .gitignore
@@ -30,3 +30,4 @@ node_modules/
 bower_components/
 client/lib
 server/api/
+server/.DS_Store
